rename Matsim forlder, better ZerodB_Source_From_Roads description, remove Building_Grid hidden input option
