+++ wps_scripts/src/main/groovy/org/noise_planet/noisemodelling/wps/Experimental_Matsim/Traffic_From_Events.groovy
@@ -13,7 +13,7 @@
  * @Author Valentin Le Bescond, Université Gustave Eiffel
  */
 
-package org.noise_planet.noisemodelling.wps.Matsim
+package org.noise_planet.noisemodelling.wps.Experimental_Matsim
 
 import geoserver.GeoServer
 import geoserver.catalog.Store
@@ -47,7 +47,6 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
 import java.sql.Connection
-import java.time.LocalDateTime
 
 title = 'Import data from Mastim output'
